feat(test.sh) add tests for 0.14 -> 0.15

+++ test.sh
@@ -1,11 +1,19 @@
 #!/usr/bin/env bash
+. ./semver.sh
 
 # kong test instance configuration
 DATABASE=postgres
-ADMIN_LISTEN=127.0.0.1:18001
-PROXY_LISTEN=127.0.0.1:18000
-ADMIN_LISTEN_SSL=127.0.0.1:18444
-PROXY_LISTEN_SSL=127.0.0.1:18443
+prefix=servroot
+export ADMIN_LISTEN=127.0.0.1:18001
+export PROXY_LISTEN=127.0.0.1:18000
+export ADMIN_LISTEN_SSL=127.0.0.1:18444
+export PROXY_LISTEN_SSL=127.0.0.1:18443
+
+prefix_2=servroot2
+export ADMIN_LISTEN_2=127.0.0.1:19001
+export PROXY_LISTEN_2=127.0.0.1:19000
+export ADMIN_LISTEN_SSL_2=127.0.0.1:19444
+export PROXY_LISTEN_SSL_2=127.0.0.1:19443
 
 POSTGRES_HOST=127.0.0.1
 POSTGRES_PORT=5432
@@ -30,12 +38,12 @@ test_suite_dir=
 ssh_key=$HOME/.ssh/id_rsa
 
 # control variables
+keep=0
 base_repo_dir=
 target_repo_dir=
 ret1=
 ret2=
 
-export KONG_PREFIX=$root/tmp/kong
 export KONG_NGINX_WORKER_PROCESSES=1
 
 # clear log file for this run
@@ -77,6 +85,9 @@ main() {
                 ;;
             -f|--force)
                 rm -rf $cache_dir
+                ;;
+            -k|--keep)
+                keep=1
                 ;;
             --ssh-key)
                 ssh_key=$2
@@ -133,7 +144,9 @@ main() {
         target_repo=$ret2
     fi
 
-    rm -rf $tmp_dir
+    if ! [[ "$keep" = "1" ]]; then
+        rm -rf $tmp_dir
+    fi
     mkdir -p $cache_dir $tmp_dir
 
     #kong prepare
@@ -145,16 +158,18 @@ main() {
     #   creates the following structure:
     #
     #   ├── cache               -> long-lived cache of git repositories
-    #   │   └── kong
+    #   │   └── kong
     #   ├── err.log             -> error logs for this run
     #   ├── test.sh
     #   └── tmp
     #       ├── kong-0.11.2     -> short-lived checkout of 0.11.2
     #       └── kong-0.12.1     -> short-lived checkout of 0.12.1
 
-    clone_or_pull_repo $base_repo
-    if [[ "$base_repo" != "$target_repo" ]]; then
-        clone_or_pull_repo $target_repo
+    if [[ "$keep" = "0" ]]; then
+        clone_or_pull_repo $base_repo
+        if [[ "$base_repo" != "$target_repo" ]]; then
+            clone_or_pull_repo $target_repo
+        fi
     fi
 
     prepare_repo $base_repo $base_version
@@ -162,6 +177,12 @@ main() {
 
     prepare_repo $target_repo $target_version
     target_repo_dir=$ret1
+
+    has_new_migrations $base_version
+    local base_has_new_migrations=$ret1
+
+    has_new_migrations $target_version
+    local target_has_new_migrations=$ret1
 
     # setup database
 
@@ -196,14 +217,27 @@ main() {
     install_kong $base_version $base_repo_dir
 
     pushd $base_repo_dir
-        set_env_vars
-        msg "Running $base_version migrations"
-        bin/kong migrations up --vv \
-            || show_error "Kong base version migration failed with: $?"
 
-        msg "Starting Kong $base_version"
+        using_kong_node 1
+
+        msg "Running $base_version migrations"
+
+        if [[ $base_has_new_migrations == 0 ]]; then
+          bin/kong migrations bootstrap --vv \
+              || show_error "Base kong migrations bootstrap failed with: $?"
+        fi
+
+        bin/kong migrations up --vv \
+            || show_error "Base kong migrations up failed with: $?"
+
+        if [[ $base_has_new_migrations == 0 ]]; then
+          bin/kong migrations finish --vv \
+              || show_error "Base kong migrations finish failed with: $?"
+        fi
+
+        msg "Starting Kong $base_version (first node)"
         bin/kong start --vv \
-            || show_error "Kong base start failed with: $?"
+            || show_error "Kong base start (first node) failed with: $?"
     popd
 
     msg "--------------------------------------------------"
@@ -217,13 +251,6 @@ main() {
         run_json_commands "before/$(basename "$file")" "$file"
     done
 
-    pushd $base_repo_dir
-        bin/kong stop --vv \
-            || show_error "failed to stop Kong with: $?"
-
-        msg "$base_version instance ready, stopping Kong"
-    popd
-
     # Install Kong target version
     install_kong $target_version $target_repo_dir
 
@@ -231,10 +258,16 @@ main() {
     # TESTS
     #######
 
-    pushd $target_repo_dir
-        set_env_vars
-        # TEST: run migrations between base and target version
+    pushd $base_repo_dir
+        using_kong_node 1
+
         bin/kong migrations list > $tmp_dir/base.migrations
+    popd
+
+    pushd $target_repo_dir
+        using_kong_node 2
+
+        # TEST: run migrations between base and target version
         msg_test "TEST migrations up: run $target_version migrations"
         bin/kong migrations up --v >&5 2>&6 \
             || failed_test "'kong migrations up' failed with: $?"
@@ -243,16 +276,48 @@ main() {
 
         $root/scripts/diff_migrations $tmp_dir/base.migrations $tmp_dir/target.migrations >&5
 
-        # TEST: start target version
-        msg_test "TEST kong start: $target_version starts (migrated)"
+        msg_test "TEST kong start (second node): $target_version starts (migrated)"
         bin/kong start --v >&5 2>&6 \
-            || failed_test "'kong start' failed with: $?"
+            || failed_test "'kong start (second node)' failed with: $?"
         msg_green "OK"
     popd
 
-    msg "--------------------------------------------------"
-    msg "Running requests against Kong $target_version"
-    msg "--------------------------------------------------"
+    if [[ $target_has_new_migrations -eq 0 ]]; then
+        msg "------------------------------------------------------"
+        msg "Running 'migrating' requests against Kong $target_version"
+        msg "------------------------------------------------------"
+
+        for file in $test_suite_dir/migrating/*.json
+        do
+            run_json_commands "migrating/$(basename "$file")" "$file"
+        done
+
+        echo
+        msg_green "*** Success ***"
+        echo
+
+        using_kong_node 1
+
+        pushd $base_repo_dir
+            unset KONG_PREFIX
+
+            bin/kong stop -p="$prefix" --vv \
+                || show_error "failed to stop Kong (first node) with: $?"
+        popd
+
+        using_kong_node 2
+
+        pushd $target_repo_dir
+            #TEST: finish pending migrations
+            bin/kong migrations finish --v >&5 2>&6 \
+              || failed_test "'kong migrations finish' failed with: $?"
+            msg_green "OK"
+        popd
+    fi
+
+    msg "------------------------------------------------------"
+    msg "Running 'after' requests against Kong $target_version"
+    msg "------------------------------------------------------"
 
     for file in $test_suite_dir/after/*.json
     do
@@ -273,6 +338,16 @@ parse_version_arg() {
     else
         ret1=$1
     fi
+}
+
+has_new_migrations() {
+  if [[ "$1" = "next" ]]; then
+    ret1=0
+    return
+  fi
+
+  semverGT $1 0.14.1 5>/dev/null
+  ret1=$?
 }
 
 clone_or_pull_repo() {
@@ -297,6 +372,11 @@ prepare_repo() {
     repo=$1
     version=$2
     tmp_repo_name=$repo-`builtin echo $version | sed 's/\//_/g'`
+    ret1=$tmp_dir/$tmp_repo_name
+
+    if [[ "$keep" = 1 ]]; then
+        return
+    fi
 
     pushd $tmp_dir
         cp -R $cache_dir/$repo $tmp_repo_name \
@@ -307,19 +387,34 @@ prepare_repo() {
         popd
     popd
 
-    ret1=$tmp_dir/$tmp_repo_name
 }
 
 set_env_vars() {
+    pref=$1
+    admin=$2
+    proxy=$3
+    admin_ssl=$4
+    proxy_ssl=$5
+
+    export KONG_PREFIX="$pref"
     if grep -q "proxy_listen_ssl" kong.conf.default; then
-        export KONG_ADMIN_LISTEN="$ADMIN_LISTEN"
-        export KONG_PROXY_LISTEN="$PROXY_LISTEN"
-        export KONG_ADMIN_LISTEN_SSL="$ADMIN_LISTEN_SSL"
-        export KONG_PROXY_LISTEN_SSL="$PROXY_LISTEN_SSL"
+        export KONG_ADMIN_LISTEN="$admin"
+        export KONG_PROXY_LISTEN="$proxy"
+        export KONG_ADMIN_LISTEN_SSL="$admin_ssl"
+        export KONG_PROXY_LISTEN_SSL="$proxy_ssl"
     else
-        export KONG_ADMIN_LISTEN="$ADMIN_LISTEN, $ADMIN_LISTEN_SSL ssl"
-        export KONG_PROXY_LISTEN="$PROXY_LISTEN, $PROXY_LISTEN_SSL ssl"
+        export KONG_ADMIN_LISTEN="$admin, $admin_ssl ssl"
+        export KONG_PROXY_LISTEN="$proxy, $proxy_ssl ssl"
     fi
+}
+
+using_kong_node() {
+  if [[ "$1" = "1" ]]
+  then
+    set_env_vars $prefix $ADMIN_LISTEN $PROXY_LISTEN $ADMIN_LISTEN_SSL $PROXY_LISTEN_SSL
+  else
+    set_env_vars $prefix_2 $ADMIN_LISTEN_2 $PROXY_LISTEN_2 $ADMIN_LISTEN_SSL_2 $PROXY_LISTEN_SSL_2
+  fi
 }
 
 install_kong() {
@@ -331,16 +426,18 @@ install_kong() {
 
     pushd $dir
         major_version=`builtin echo $version | sed 's/\.[0-9]*$//g'`
-        if [[ -f "$root/patches/kong-$version-no_openresty_version_check.patch" ]]; then
-            msg "Applying kong-$version-no_openresty_version_check patch to Kong $version"
-            patch -p1 < $root/patches/kong-$version-no_openresty_version_check.patch \
-                || show_error "failed to apply patch: $?"
-        elif [[ -f "$root/patches/kong-$major_version-no_openresty_version_check.patch" ]]; then
-            msg "Applying kong-$major_version-no_openresty_version_check patch to Kong $version"
-            patch -p1 < $root/patches/kong-$major_version-no_openresty_version_check.patch \
-                || show_error "failed to apply patch: $?"
-        else
-            msg "No kong-no_openresty_version_check patch to apply to Kong $version"
+        if [[ "$keep" = 0 ]]; then
+          if [[ -f "$root/patches/kong-$version-no_openresty_version_check.patch" ]]; then
+              msg "Applying kong-$version-no_openresty_version_check patch to Kong $version"
+              patch -p1 < $root/patches/kong-$version-no_openresty_version_check.patch \
+                  || show_error "failed to apply patch: $?"
+          elif [[ -f "$root/patches/kong-$major_version-no_openresty_version_check.patch" ]]; then
+              msg "Applying kong-$major_version-no_openresty_version_check patch to Kong $version"
+              patch -p1 < $root/patches/kong-$major_version-no_openresty_version_check.patch \
+                  || show_error "failed to apply patch: $?"
+          else
+              msg "No kong-no_openresty_version_check patch to apply to Kong $version"
+          fi
         fi
 
         msg "Installing Kong..."
@@ -363,10 +460,6 @@ run_json_commands() {
     if [[ -f $filepath ]]; then
       resty -e "package.path = package.path .. ';' .. '$root/?.lua'" \
             $root/util/json_commands_runner.lua \
-            http://$ADMIN_LISTEN \
-            http://$PROXY_LISTEN \
-            http://$ADMIN_LISTEN_SSL \
-            http://$PROXY_LISTEN_SSL \
             $filepath >&5 \
             || failed_test "$name json commands failed with: $?"
         msg_green "OK"
@@ -376,7 +469,10 @@ run_json_commands() {
 }
 
 cleanup() {
-    kill `cat $KONG_PREFIX/pids/nginx.pid 2>/dev/null` 2>/dev/null
+    kill `cat $base_repo_dir/$prefix/pids/nginx.pid 2>/dev/null` 2>/dev/null
+    kill `cat $base_repo_dir/$prefix_2/pids/nginx.pid 2>/dev/null` 2>/dev/null
+    kill `cat $target_repo_dir/$prefix/pids/nginx.pid 2>/dev/null` 2>/dev/null
+    kill `cat $target_repo_dir/$prefix_2/pids/nginx.pid 2>/dev/null` 2>/dev/null
 }
 
 show_help() {
